Implementing compare images from local files and USB drive

+++ Assets/MyScripts/RecordInputs.cs
@@ -24,14 +24,14 @@ public class RecordInputs : MonoBehaviour{
     private string JUMP = "_";
     private string DASH = "_";
     private string ATTACK = "_";
-    private string PAUSE = "_";
+    //private string PAUSE = "_";
 
     //X and Y position of the player
     private float Xpos = 0f;
     private float Ypos = 0f;
 
     private double nextActionTime = 0.0;
-    public double peroid = 0.4;
+    public double peroid = 0.1;
 
     int count;
     byte[] byteArray;
@@ -66,7 +66,7 @@ public class RecordInputs : MonoBehaviour{
         JUMP = "_";
         DASH = "_";
         ATTACK = "_";
-        PAUSE = "_";
+        //PAUSE = "_";
 
         //reset the Pos
         Xpos = 0f;
@@ -124,11 +124,11 @@ public class RecordInputs : MonoBehaviour{
             //record that attack was pressed this frame
             ATTACK = "Attack";
         }
-        if (CrossPlatformInputManager.GetButtonDown("Pause"))
-        {
-            //record the game was paused
-            PAUSE = "Pause";
-        }
+        //if (CrossPlatformInputManager.GetButtonDown("Pause"))
+        //{
+        //    //record the game was paused
+        //    PAUSE = "Pause";
+        //}
 
         //if x seconds have passed
         //then get the current player position and write it to a file
@@ -141,7 +141,7 @@ public class RecordInputs : MonoBehaviour{
         //write the current values at the end of every frame
         //if there has been a change
         //if (X != 0f || Y != 0f || JUMP != "_" || DASH != "_" || ATTACK != "_")
-        WriteInputs(X, Y, JUMP, DASH, ATTACK, PAUSE, Xpos, Ypos);
+        WriteInputs(X, Y, JUMP, DASH, ATTACK, Xpos, Ypos);
 
         //Debug.Log(destinationDrive);
     }
@@ -153,12 +153,12 @@ public class RecordInputs : MonoBehaviour{
 
     }
 
-    void WriteInputs(float w_X, float w_Y, string w_Jump, string w_Dash, string w_Attack, string w_Pause, double w_XPos, double w_YPos)
+    void WriteInputs(float w_X, float w_Y, string w_Jump, string w_Dash, string w_Attack, double w_XPos, double w_YPos)
     {
         int frames = Time.frameCount;
         frames--;//adjust for list starting at 0
         //combine all values for this frame into one string
-        string input = w_X + "," + w_Y + "," + w_Jump + "," + w_Dash + "," + w_Attack + "," + "w_Pause" + "," + Time.deltaTime +"," + w_XPos + "," + w_YPos + ",";
+        string input = w_X + "," + w_Y + "," + w_Jump + "," + w_Dash + "," + w_Attack + "," + Time.deltaTime +"," + w_XPos + "," + w_YPos + ",";
 
 
         //create writer object & write to file
@@ -174,8 +174,15 @@ public class RecordInputs : MonoBehaviour{
 
     private void OnApplicationQuit()
     {
-        //move the file to the external disk
-        File.Move(file, destinationDrive);
+        if (!File.Exists(file))
+        {
+            Debug.Log("Extral Drive not connected. Could not move file");
+        } else
+        {
+            //move the file to the external disk
+            File.Move(file, destinationDrive);
+        }
+        
     }
